Enable /W3 and fix /W3 warnings for Windows
Replace most functions deemed unsafe by the compiler with the more
secure CRT functions:

 - gmtime() becomes gmtime_s(), or gmtime_r() for Linux builds

 - getenv() becomes getenv_s()

 - sscanf() becomes sscanf_s()

 - sprintf() becomes sprintf_s()

 - fopen() becomes fopen_s()

 - strcpy() becomes strcpy_s()

 - strcat() becomes strcat_s()

 - strncat() becomes strncat_s()

Also use more explicit conversions to prevent type and pointer
mismatches. Move most #ifdef branches to the following files:

 - include/openenclave/host.h

 - include/openenclave/internal/datetime.h

 - tests/crypto/readfile.h

Fix part of #2075.

+++ tests/mbed/enc/enc.c
@@ -149,7 +149,7 @@ static oe_result_t _syscall_hook(
         case SYS_lseek:
         {
             int rval = 0;
-            result = mbed_test_lseek(&rval, (int)arg1, (off_t)arg2, (int)arg3);
+            result = mbed_test_lseek(&rval, (int)arg1, (int)arg2, (int)arg3);
             break;
         }
         case SYS_readv:
